+++ server/src/main/java/org/allaymc/server/eventbus/event/network/PacketEvent.java
@@ -1,10 +1,11 @@
-package org.allaymc.api.eventbus.event.network;
+package org.allaymc.server.eventbus.event.network;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.eventbus.event.CancellableEvent;
+import org.allaymc.api.eventbus.event.network.NetworkEvent;
 import org.cloudburstmc.protocol.bedrock.packet.BedrockPacket;
 
 /**