+++ server/src/main/java/org/allaymc/server/eventbus/event/network/PacketSendEvent.java
@@ -1,4 +1,4 @@
-package org.allaymc.api.eventbus.event.network;
+package org.allaymc.server.eventbus.event.network;
 
 import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.cloudburstmc.protocol.bedrock.packet.BedrockPacket;
@@ -6,8 +6,8 @@ import org.cloudburstmc.protocol.bedrock.packet.BedrockPacket;
 /**
  * @author daoge_cmd
  */
-public class PacketReceiveEvent extends PacketEvent {
-    public PacketReceiveEvent(EntityPlayer player, BedrockPacket packet) {
+public class PacketSendEvent extends PacketEvent {
+    public PacketSendEvent(EntityPlayer player, BedrockPacket packet) {
         super(player, packet);
     }
 }